Move internal/mounttree to daemon/internal/mounttree
