+++ daemon/containerfs_linux.go
@@ -17,7 +17,7 @@ import (
 
 	containertypes "github.com/docker/docker/api/types/container"
 	"github.com/docker/docker/daemon/container"
-	"github.com/docker/docker/internal/mounttree"
+	"github.com/docker/docker/daemon/internal/mounttree"
 	"github.com/docker/docker/internal/unshare"
 	"github.com/docker/docker/pkg/fileutils"
 )
